changes national id to long so it can only be 14 characters long . check profile picture extension so it can only be .jpg / .png . change Customer area to User area

+++ Models/ApplicationUser.cs
@@ -11,7 +11,7 @@ namespace Models
     {
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public int NationalId { get; set; }
+        public long NationalId { get; set; }
         public string Address { get; set; }
         public string? ProfilePicture {  get; set; }
